Reorganice JS
